Fixing earpiece amplification in call.
HTC released a non-finished version of their kernel
that doesnt contain all the RCV amp parts.

Added a headset_amp call to htc_headset_mgr part,
so it can call the soundcard of m7 directly when
headset is plugged/unplugged.

Based on that states and the SPKR amp events
we use the GPIO commands to enable/disable
RCV part (earpiece amp) of the phone.

Thus earpiece audio is working while in call,
and switches on/off based on headset state
as well.

Signed-off-by: Bryan Smith <zarboz@gmail.com>